mauve: rework tests, exclusions and test setup
Regenerated all the blacklists based on a successful run using Java 6
on x86-64.  As a result the blacklists are considerably shorter which
is less likely to have an impact on command line (shell) length when
invoking the tests.

The mauve git repo that contains the pre-built class files must now be
pre-built within /tmp/mauve or no tests will be run.

The parser has been removed that checks for pass, fail or skip as the
Mauve test harness, combined with Lava, will produce one row per test
which results in approximately 250 pages in the Lava dashboard.
Instead you now get a high-level pass/fail for each top-level Java
package that is tested.

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>