mauve: use modified version of the test runner for LAVA
The current version of the test harness has too many issues with:

  spawning many threads (5000+)
  timeout doesn't work properly
  timeout leaves the pipe process in a hung state
  VM size grows to 13+ GB
  eventual stack overflow when restarting after a timeout

and as such the tests do not run to completion on Open Embedded
aarch64 using the ARM fastmodel.

Rather than trying to fix these issues in the existing Harness I have
duplicated the file (now HarnessLinaro.java) where all of the issues
listed above now seem to be fixed.  However, given that this hasn't
had much soak testing (1 day) and I really want to get the LAVA tests
to run to completion, we will need to revisit these changes to see if
they are suitable for pushing upstream.

This commit now invokes the tests using HarnessLinaro[.java].

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>