jtreg: accomodate differences between jdk and hotspot tests
Also attach ARCHIVE_BUNDLE.zip, if available.  It also turns out that if
you test jdk_beans then the underlying JDK makefiles actually tests
jdk_beans{1,2} and {1,2} become the directory names and, as a result,
the test results don't get appended correctly.  Updated the script to
glob on the test name so that we don't have to peak too much into the
JDK test machinations.

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>