jtreg: extract and attach summary results for hotspot tests
When a hotspot test completes the underlying JDK hotspot makefile
deletes the intermediate files leaving only the ARCHIVE_BUNDLE.zip to be
atached as a result of the test.  Unfortunately this makes it hard to
see at a glance what the outcome of the test is when browsing through
the results in LAVA. This commit unzips JTreport/text/*.txt and attaches
them as individual files for convenience.

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>