hackbench: update measurements
modified the hackbench test suite to run a fixed number of iterations
and added a script called dd_infinite.sh that does the dd part. This is
useful for having relevant measurements, since in idle the performance is
usually good. The results are gathered by hackbench_parse.py which outputs
4 measurements, minimum value, maximum, average and median.

tested on validation.linaro.org, results can be found at
http://validation.linaro.org/scheduler/job/71437

Signed-off-by: Ciprian Barbu <ciprian.barbu@linaro.org>
Acked-by: Fathi Boudra <fathi.boudra@linaro.org>