Rework OpenJDK tests to invoke jtreg directly
This commit reworks the invocation of the JTREG tests to rely on
jtreg(1) directly. It was previously using the makefiles in the OpenJDK
test directories but there were significant differences between the JDK
Makefile and the hotspot Makefile which added unnecessary complexity,
particularly as we now want to run tests against both -client and
-server VMs.

There are also additional parameters for specifying the test timeout and
job concurrency.  And I also moved the exclude file into this repo so
that it can be subject to code reviews.

Change-Id: Ic90df7969f862c3fc5ecb71f658601cd2e1d89ba
Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>