jtreg: add failed results as discrete attachments
For each test that fails add the corresponding .jtr (jtreg result) for
easier observation from the LAVA dashboard.  Also attach a zip of the
JTwork and JTreport directory sans the class files.  This zip is now
attached directly to the overall LAVA test using lava-test-run-attach.
The overall execution is now run using jtreg and the results are then
parsed and analysed for success/fail.

Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>
Change-Id: Id5e918e196892e6935b3361f822f6f129f59772c