Revert "add openem x86 tests v4"
This reverts commit 82eab5d612ef245066fc475811fae082e93931eb.

I was a bit too fast with merging that patch. There are still comments that were not addressed as Fathi pointed:
1. about location of DPDK on people.linaro.org:
could you add a comment/justification on why you do that? why we don't pull for website and have to carry a copy? does the license permit to redistribute this way?

2. about dependencies:
 - are all the deps needed?
 - definitely, no. please clean up the build deps.

Change-Id: I0a1b0d50a32baeac6856f9b95905b50e2b12db10