is-cpu-isolated: Use Tick count of last loop
We are calculating time interval between ticks here. Currently it happens this
way:
1. Start Count
2. End Count
3. Take time difference between two
4. Repeat above three again.

The problem here is that we do a Start Count almost just after End Count. And
this can take a lot of time, i.e. equal to the time CPU is isolated. Which is
around 40 seconds for now.

So change this logic to do following:
1. Start Count
2. End Count
3. Take time difference between two
4. Start Time = End Time
5. Repeat 2->4 again.

Change-Id: I455c9d5182c2ca43544848a6524a61580fe56cc4
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>