Update patch set 4
Patch Set 4:

Tyler, I have tried the test on both ubuntu and OE images before. It appears to run well on ubuntu , but I have not been successful with OE yet. lshw seems to be missing in the images that I tried to run the test on. Milosz said for OE it has to come pre-installed in the images since it is not recommended to add the install deps steps in the yaml file for OE tests.

Patch-set: 4
Reviewer: Esla Anzaku <1000056@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0