new: add qemu build and LTP tests
This introduces two new scripts to aid with build and test of qemu. This
first (qemu-binfmt.sh) is a cut down version of the upstream binfmt
script for setting up qemu-${ARCH} as a binfmt helper. The
second (qemu-ltp.sh) is because we need to execute the LTP build inside
the chroot and this seemed neater than adding a chroot to every .yaml
line.

The actual test definitions are split between:

  - qemu-linux-user-build.yaml - build and install
  - install-qemu-chroot - installing chroots for test
  - qemu-ltp.yaml - running the test

Change-Id: I4b0571f845d73da190c1362e43bb2fe1e323c0f3
Signed-off-by: Alex Bennée <alex.bennee@linaro.org>