Update patch set 1
Patch Set 1:

would it be possible to make qemu, ubuntu and openemedded use the same script in the run step?
and maybe rework so ubuntu and qemu are using the same install script?
for the parsing is there some reason to TBROK, TCONF and so on, and not use the same output (results) like openembedded does (in the results directory)?

Patch-set: 1
Label: Code-Review=0