new: add qemu-system-test definition and expect scripts
As we need to send commands to the qemu console to shutdown cleanly and
it's a pain to custom build a rootfs to just run a test we use an expect
wrapper for controlling qemu. The whole thing is controlled by a wrapper
so several TESTS can be run in one definition and the results kept
together.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Change-Id: I82ae7305ce2cdd8d3cef05621f4e4eace7d35bb7