new: add qemu build and LTP tests
This introduces two test definitions to aid with build and test of qemu.
They are:

  - build-qemu-from-source.yaml - build and install
  - install-qemu-chroot - installing chroots for test

There is also a change to make the ubuntu LTP tests use a script for
running so the whole thing can be wrapped in a chroot and run under QEMU
control. As we need the HEAD of the LTP tests for aarch64 currently
added a mode to build against the latest LTP tests

Change-Id: I4b0571f845d73da190c1362e43bb2fe1e323c0f3
Signed-off-by: Alex Bennée <alex.bennee@linaro.org>