is-cpu-isolated.sh: Enable tracking isolation for system with more than 2 CPUs
Till now this scripts was only useful for a system with two CPUs as there were
some assumptions. But there were requirements from multiple people about
supporting more platforms which might have more than two cores.

Hence this patch, it enables support for multiple cores and any core can be
isolated (leaving core 0, which can't be isolated in kernel as that's the boot
CPU).

Order of parameters is also changed based on what user may want. And so
nohz-isolation.yaml is updated in the same patch so that things don't break
inbetween.

Change-Id: I0656131a6a6d043e662d16f05dd08f172f2a3d9d
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>