is-cpu-isolated: Fix incorrect list of non-isol-cpus
update_non_isol_cpus() was incorrectly expecting ISOL_CPUS to contain only one
CPU and would fail badly if a comma separated list of CPUs is passed to it.

For example:
run ./is_isol_cpus 1,2 [out of 4 physical cpu, cpu 1 & 2 to isolate]
would result in: Isolate: CPU 1,2 and leave others: 0,1,2,3

Fix it by adding another routine to find if a cpu is isolated one or not.

Change-Id: I64e853dc55efd8fb668e2a3f628e5497aa6d009f
Signed-off-by: santosh shukla <santosh.shukla@linaro.org>