android: added lava-android-test wrapper
The lava-test-shell wrapper around lava-android-test
allows to execute the lava-android-test from the KVM device
type in LAVA which acts as a host. With this approach
it is possible to controll the host environment and for example
use different Java versions with different Android CTS branches.

The test wrapper JSON fragment follows:
{
    "command": "lava_test_shell",
    "parameters": {
        "role": "host",
        "testdef_repos": [
            {
                "git-repo": "git://git.linaro.org/qa/test-definitions.git",
                "parameters": {
                    "TEST_NAME": "cts",
                    "TEST_PARAMS": "--package android.bionic --timeout 1800 --disable-reboot"
                },
                "testdef": "android/lava-android-test-host.yaml"
            }
        ],
        "timeout": 7200
    }
},
{
    "command": "lava_test_shell",
    "parameters": {
        "role": "target",
        "testdef_repos": [
            {
                "git-repo": "git://git.linaro.org/qa/test-definitions.git",
                "parameters": {
                    "TEST_NAME": "cts",
                    "TEST_PARAMS": "--package android.bionic --timeout 1800 --disable-reboot"
                },
                "testdef": "android/lava-android-test-target.yaml"
            }
        ],
        "timeout": 7200
    }
}

lava-android-test-host.yaml and lava-android-test-target.yaml take
TEST_NAME and TEST_PARAMS. These have to be the same for each pair
of test shells. Optionally lava-android-test-host.yaml takes
JAVA_PACKAGE parameter which is used to define the Java version
running on host.

Change-Id: Ief169010a1510392c741513136cbc245ab769b7b
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>