vm: add ltp test which is used by vmgroups
When vmgroups machenism is introduced, any test in a guest vm doesn't need the
original method common/scripts/kvm/test-kvm.sh any longer, in the json file,
submitting tests for vm is just as same as for host, some test scripts for host
can be reused in vm -- only if the necessary environments and tools etc are
there as in the host, if this condition isn't met, e.g. something is missing in
vm, then we need additional steps to setup it.
To achieve this, the vm/ directory is intoduced to hold all these corresponding
files for this purpose.
And this patch adds scripts for ltp test in vm, then vmgroups style json can
use this directly, just as same way as host test.
Other scripts can be created and put into this vm/ later.

Change-Id: I2a2b8c92ea1bc4e5fba74e24ceb1b346c7c9cb2d
Signed-off-by: Hongbo Zhang <hongbo.zhang@linaro.org>