Update patch set 1
Patch Set 1:

Yes I copied your script, and I didn't care which version of hackbench to be used.
There is a problem with using the current openbedded/hackbench.yaml, the hackbench isn't built in the vm image:
https://validation.linaro.org/scheduler/job/144122/log_file
so we need to do extra installation for vm, this can be done by:
a) introduce a specific script for vm as this patch
b) rebuild vm image, this is not reasonable I think.
c) edit the current openbedded/hackbench.yaml to work.

You choose c I can understand, I choose a because:
 I think openembedd and ubuntu is rich with many tools build in, we may likely lack such tools in vm, then if we do more tests in vm  in future, maybe the current scripts for openembedded and ubuntu don't work and maybe it is hard to patch them to work.

Again, I understand your point, the ubuntu/ltp.yaml works for vm, but the openbedded/hackbench.yaml doesn't, if we drop patch like this review, we still need a patch for current openbedded/hackbench.yaml.

Patch-set: 1