Update patch set 1
Patch Set 1:

Hongbo,

I think you're missing the point here. I'm perfectly fine with 2 versions of tests:

* openembedded (assumes everything is built into the rootfs)

* ubuntu (installs stuff on the target before running the tests)

If there is no 'hackbench for Ubuntu' YAML, I'm happy to accept the patch. On Ubuntu hackbench is a part of rt-tests package. So you don't have to rebuild anything. Ingo's version of hackbench is different and is targetting KVM benchmarking (only?). 

Which rootfs are you using for guest?

Patch-set: 1