Update patch set 1
Patch Set 1:

OK, now we find the ubuntu/rt-hackbench.yaml, and there is the installation we need in this script, so we use this script in vm test.
But here comes a further problem, the git and gcc for downloading and compiling are missing in the vm again.
There are two solutions for this issue:
the 1st is adding git and gcc into the 'deps' list, it works as I've tested.
the 2nd is pre-build the git, gcc and also wget into vm image, this way is better I think for this time because these tools are the most basic ones for tests.

I was using the vm image for snapshot:
http://snapshots.linaro.org/ubuntu/images/kvm-guest/14/

Patch-set: 1
Reviewer: Hongbo Zhang <1000086@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0