Update patch set 3
Patch Set 3:

It 'should' run on TC2 with LSK 3.14. I think the configs are enabled there. I submitted a job, but it's going to take a while before it returns anything (the queue is long):
https://validation.linaro.org/scheduler/job/175284
I'll try to push the LAB guys to make the staging board online so I can actually try there.

Patch-set: 3