Update patch set 3
Patch Set 3:

On this patch set 3, I restored the path to /system/bin, however I'm going to submit another patch set to delete this test, as Naresh suggested. So please pick up this patch set 3 as the final one to apply the changes to scripts inside the android build image.

Patch-set: 3
Label: Code-Review=0