Update patch set 2
Patch Set 2:

For the failures of glslparser test, I will check that with others.
if we decide to remove those test cases, I will update on the piglit source side, should be no change on this test definition side.

And about the copyright declaration, is there any rules to select?
For these test, the same license as others should be OK I think.

Patch-set: 2
Label: Code-Review=0