Update patch set 1
Patch Set 1:

The variable number of results is hard to track. We used to have this kind of situation with pm-qa. The number of results jumped from job to job between ~20 and ~100. So it was really hard to tell what was going on. But as you mentioned there is no other easy way of doing this test with measurements.

As for read vs write my concern is the test name. Can we name it just read_perf?

Patch-set: 1
Label: Code-Review=0