Update patch set 4
Patch Set 4:

Another idea. Since we would like it to run on both android and non-android, the shebang line will be different. So maybe enclose the test in the function, create 2 separate short scripts (1 for android, 1 for ubuntu/oe) and run it this way?

Please change tabs to spaces.

Patch-set: 4
Label: Code-Review=0