Update patch set 1
Patch Set 1:

except the method for getting the apk file and the methods about getting streamline data(which is not used in the apk automation test I think), I don't think there is much in common.  From the point of the structure of the programming, it's a little similar(prepare-start-run-post), but the logic is different I think and different files need to be parsed. If you want to make a common framework for this application test and that benchmark automation test, I think it's better to plan another (maybe 1w)work to do that.  And about where to put this test, I don't have much comment. put it here just remembered that you said android-apk-automation.git should store test which uses ViewerClient.

Patch-set: 1