Update patch set 1
Patch Set 1:

Can we schedule a hangout on this one? I think it should go to android-apk-automation. It's not tied with AVC. The framework is flexible enough to work with any script. What the framework does is:
 * download apk
 * install apk
 * run user script
 * uninstall apk
So any script can be run. And as you mentioned the gator stuff is already included, which means a bit less code to maintain.

Patch-set: 1