Update patch set 1
Patch Set 1:

Tested in lava, 
https://validation.linaro.org/dashboard/streams/anonymous/chase-qi/bundles/663ceafcf8bdd61083904de88d1f08c8e7520947/

Most of the time, 64bit ION test failed to finish, keep outputting warnings like this https://validation.linaro.org/scheduler/job/252254/log_file#L_1701_9812, until timed out. I guess it is a issue of test source code or a bug of the kernel. It should not be a issue of testcase.

Patch-set: 1