Update patch set 7
Patch Set 7:

Sorry, MIlosz, I missed your comments:( I made too much commits on this change before test in case I deleted some changes locally).

For this change, the target is mainly to get the run time for each test, and use the run time as the benchmark data to check the bionic performance. If the gtest wrapper can do this as well, that would be greate.

Patch-set: 7