Update patch set 2
Patch Set 2:

Arthur,
What I meant was that you can't distinguish between pmqtest and ptsematest in the result bundle. The difference is hidden in the test shell parameters. So I'd rather see:
pmqtest#1 -> #0 (Min)
ptsematest#1 -> #0 (Min)

and so on (or sth similar to that). This way it will be much easier to plot stuff in LAVA (and potentially somewhere else) directly from the bundle data.

I hope it's clear now.

Patch-set: 2
Label: Code-Review=0