ubuntu: ltp: fix out of disk space issue
The current approach is download and install LTP in test folder, which
causes the size of the folder is very big. Once the compressed size of
the folder exceed the disk space available on master image, it will
lead to result collection failure.

Alternatively, we can download and install LTP under /opt/ltp directory.
In this way, it addresses the out of disk space issue, and we only need
compile and install LTP once for multiple LTP lava_test_shells in the
same LAVA job.

Change-Id: I5f5c37c38c2cbabb77c1a2227c5336a93cf872f2
Signed-off-by: Chase Qi <chase.qi@linaro.org>