Android: CTS: modified multinode CTS
Multinode CTS was modified in order to gracefully handle device
disconnection events and save logs even in exeptional situations.
Results are presented as number of pass/failed/executed for each
package. All log files are attached to the results. This should bring
smaller log files in LAVA and more readability to the results.

Change-Id: Idd2afc4fea63f2ea66e07ad4e4608ed7401b95f6
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>