To add the PLATFORM variable required for ltp-ddt
LTP-DDT requires the platform to be specified because some testsuites
have platform dependencies. Also, the LTP-DDT prints logs to stdout as
well as to a log file if a log file is specified in the arguments. To
avoid the duplication of the parsing of test results, using cat on the
log is avoided for LTP-DDT

Change-Id: Ic630dd08444e0391ce163efbd7ea9129a63838e4