Update patch set 1
Patch Set 1:

(2 comments)

This is a reference to resolve for the aapt problem with cts6.0 package.

http://people.linaro.org/~yongqin.liu/cts/android-sdk-linaro.tgz is the package I created which after updated with the android command from google SDK.
We might be able to use a package like the adb and fastboot we used before, but I don't know much about that, so this change using adb/aapt from Android SDK is only reference. we can update this change after we decided the right direction.

Patch-set: 1
Label: Code-Review=0