Update patch set 1
Patch Set 1:

I understand the concerns that have been highlighted. My thought on the best way to go about this is to have a different yaml file for the ltp-ddt project. This project is substantial project on its own and although compatible with the ltp, there are modifications that can cause conflicts. 
Also, maybe most people that run ltp-ddt might not run ltp tests and the other way round. To avoid conflicts and confusion, I think having a different yaml and script for the ltp-ddt might be the preferred option.

Patch-set: 1