ubuntu: ltp: fix parse pattern
ltp-pan print test case, result and exitcode in the format "%-30.30s
%-10.10s %-5d\n". Most of the time, exitcode doesn't take all the 5
digital placeholder, in this case, there are spaces after the exitcode.
That is why the current pattern doesn't work with the new
lava-local-run I created.

The current pattern actually works in LAVA instance, it is not able to
match skipped tests though. Further investigation shows that the
additional spaces are stripped off in LAVA's output for some reasons
that I don't know. Refer to
https://validation.linaro.org/dashboard/attachment/3111331/view

The new pattern in this patch fixed this issue and also able to match
skipped tests. LAVA job example
https://validation.linaro.org/scheduler/job/1025927

Change-Id: Iafcbd11a0bc3588eba51cafef53ce9082ab9aa36
Signed-off-by: Chase Qi <chase.qi@linaro.org>