Update patch set 1
Patch Set 1:

(3 comments)

Please refer to the commit msg for the basic ideas on the refactoring. It is a new start, we probably should consider all the details and plan for the long term.

Please take your time and review it as deeply as possible, and let me if you have any comments.

In the long term, we probably will need to write more docs to say how to run a test and how to create a new one.

Patch-set: 1