Update patch set 1
Patch Set 1:

Without this change all the benchmark(multinode) tests will fail for Nougat builds.

But event with this change,the benchmark test might fail too because ui changed.

So please check and merge this change first, and then we will know which benchmark tests will fail because of the UI changes.

Patch-set: 1