httperf-client: client side of http performance tests
Script and YAML file to setup the host to run httperf against an http
server. The tests measure the performance of the server depending on the
number of cores.

This script is meant to run in a multi-node configuration, where it will
coordinate with a server side of the test in order to determine when the
tests can start, and when they are done.

Change-Id: I3af563086a8fba64f1b0f32c0878761997d21158
Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>