httperf-client.sh: client side of http performance
This script needs to be run in LAVA, it will setup the host to run
httperf against an http server.
This script is meant to run in a multi-node configuration, where it will
coordinate with a server side of the test in order to determine when the
tests can start, and when they are done.

Change-Id: I3af563086a8fba64f1b0f32c0878761997d21158
Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>