httperf-nginx-server.sh: server side of the http tests
This script should be run in LAVA, it coordinates with httperf-client.sh
to initiate and proceed with more tests.
Currently it supports testing NGiNX on plain linux IP stack, and a
special version of NGiNX compiled for OFP and ODP-DPDK, however the
latter is not too portable at the moment and will require further work.
Whether one or the other is tested is configured by the CONFIG_TYPE
environment variable: linux-ip or odp-dpdk, respectively.

This relies completely on VLAND and lava-{send,wait}.

Change-Id: I775083f7183895039064141b51172a3280249848
Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>