nginx-server: server side of the http performance tests
The nginx-server.sh script implements the same protocol as the
httperf-client.sh script, namely:

client                           server
   |                                |
   |<----------- num_cores ---------| # server indicates num cores
   |---------- client_ready ------->| # server is configured for 1 core
   |<-- server_num_cores_1_ready ---| # client runs tests for 1 core
   |--- client_num_cores_1_done --->| # client is finished testing
   |                                | # server reconfigures for 2 cores
   |<-- server_num_cores_2_ready ---| # client runs tests for 2 cores
   |--- client_num_cores_2_done --->| # client is finished testing
   |                                | # repeated until num_cores in
   |                                | # increments of 2 (1, 2, 4, 6...)
   ...

Currently it supports testing NGiNX on plain linux IP stack, and a
special version of NGiNX compiled for OFP and ODP-DPDK, however the
latter is not too portable at the moment and will require further work.
Whether one or the other is tested is configured by the CONFIG_TYPE
environment variable in the test definition: linux-ip or odp-dpdk,
respectively.
Also the maximum amount of cores for which to test can be limited by
setting the MAX_CORES environment variable (0 meaning all cores).

This relies completely on VLAND and lava-{send,wait} for
synchronization.

Change-Id: I775083f7183895039064141b51172a3280249848
Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>