Update patch set 4
Patch Set 4:

That is probably the easy (only?) way. It should be done in the test script so if execution is attempted outside LAVA, the script just exits doing nothing.

As for this patch there is not much to do. The scripts will just fail if executed outside LAVA. I'm not sure if it makes sense to do the lava check already here.

Patch-set: 4
Label: Code-Review=0