Update patch set 2
Patch Set 2:

Doesn't work in docker as there is no systemd, but tested manually and works well. I'm wondering whether the apache-bench stuff should be moved to utils? It's used in 2 tests now. Or maybe the apache-bench test should be customized to run apache or nginx or tomcat. I think the latter is better.

Patch-set: 2
Label: Code-Review=0