Update patch set 3
Patch Set 3:

> Looks OK with exception of commit message which should be a bit
 > more verbose. Also origin of dsdbench should be mentioned somewhere
 > (README?).

More information added to commit msg and test description. In this test, we are not including binaries here, we clone the repo, compile and run tests. Added two comment lines in the script to say what is the test for and where is the source.

Patch-set: 3
Reviewer: Chase Qi <1000096@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0