android boottime2: collect time from init start to surfaceflinger start
CONSOLE_SECONDS is the time from kernel start to init started,
and the TIME_VALUE is reported by surfaceflinger service
which is the time from surfaceflinger started to UI booted up.

There is still the time from point init started to the point
service surfaceflinger started point not collected.

This change is to collect that time, and add it into the total time.

Change-Id: I4dbd8a3ca4d2afb536b65cc9bc245ce30dc06b99
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>