Update patch set 6
Patch Set 6: Code-Review+2

That looks good. I had a brief chat with Anders on IRC. He suggested that all scripts should be copied to test-definitions in order to remove dependency on git (and avoid creating OE recipe). Please check with Anders about details. I'm fine merging the patch as it is now. If you decide to add the scripts from LNG repo here, I'm fine with that as well.

Patch-set: 6