Update patch set 9
Patch Set 9:

Ravi,
For me it looks good (as I wrote for v6). If you agreed with Anders that it can be merged and the isolation script remains in the LNG repo, I'm fine with that. Just let me know. 

Sorry for delay, I'm working on ERP release now and this takes all my cycles.

Patch-set: 9
Label: Code-Review=0