automated: sh-test-lib: do not source /etc/lsb-release
If /etc/lsb-release file exists , then it will be sourced in dist_name() so any
variable defined in this file will override/pollute our existing
environment. Since dist_name() is called in the context of the test script, we
might end up polluting our test environment with unneeded variable, or even
worse we might override variables that we have set in our test scripts...

Instead of sourcing the file in the current env, let's do it in a subshell.

Change-Id: I99ff9b7bb498f47f01adfe0d21f9b1cce0ebf2cc
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>