Update patch set 1
Patch Set 1:

It works if you skip install step. It lsb-release is not there and you don't skip install, that won't work.

On the other hand we're missing the following dependencies in the script:
coreutils
procps
util-linux
iproute2

The question is whether there can be a minimal distro without any of the above?

Alternatively the check_root should be moved inside install function so it's only needed if we attempt to install something.

Patch-set: 1