automated: test-runner: encapsulate the TestRun object into TestDef
There is no need to create an additional object in main(). Instead we can create
the 'runner' object in the TestDef init() and directly invoke the run() and
check_result() function.

Also rename TestDefinition.run() function as TestDefinition.mkrun(). This
function really generates the run script, calling it run() is confusing, as this
is not where the test runs.

Change-Id: I4348e1e834e7083174aa60a8f4533747e49c3757
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>