nginx-server: remove dependency on private repo
To work around some limitation on the current port of NGiNX to OFP,
a private repo with specially built OFP and NGiNX was used by the test
script. This is not desirable in the long run, so we will resort to
installing the packages provided by LNG at deb.opendataplane.org, and
add the limitation that NGiNX for OFP will only be tested with one
worker for now.

When support for more workers is available, the limitation should be
removed.

Change-Id: I0f2d5f1c4efe9efb53c77270c0bebaeb03c709b6
Signed-off-by: Josep Puigdemont <josep.puigdemont@linaro.org>