test-runner: add remote execution feature
Test can be executed remotely using ssh. It is assumed that ssh
communication is performed without passwords and public key is already
added to authorized_keys.

Most tests require root access to be executed, so there is a silent
assumption that remote test execution is done using 'root' account. The
account name might be different, but should be granted the same level of
privileges.

Change-Id: I41f1d32d20f15ac6cfa8fc9e99f527fe61d4743d
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>