Update patch set 2
Patch Set 2:

(8 comments)

hi, this is a first review for now.. i will continue to look into that but I don't want to loose the comments made so far, so i am posting it..

I was able to run a single test and also a plan. i have to say, that it works really well.. there are a few quirks but nothing that seems major. 

and it is very convenient to run the tests from the PC ;-)

for reference, this is the test plan i am using for now
https://git.linaro.org/people/nicolas.dechesne/test-definitions.git/tree/plans/db410c/linux.yaml?h=dev

Patch-set: 2