Update patch set 2
Patch Set 2:

> I don't like this solution/idea. Change from measurement to
 > pass/fail is sth I would like to avoid.  Is this really so
 > disturbing?

I think the problem here is that we should not have measurement for failed test cases.

And in a chart, if the range from maximum and the minimum(the failed negative value) is very large, then the chart is not hat easy to be read.
And test cases without measurement would could help to exclude such bad case for the chart I think.

BTW, maybe the chart should be changed to only select values from passed test cases.

Patch-set: 2