Update patch set 2
Patch Set 2:

I'm not sure which chart you're referring to, but for me this is a major problem. Either the test result is a measurement (which btw should not be associated with pass/fail value) or it's a pass/fail result. It's very inconvenient to alter between these basing on measured value. I don't think it's a good idea to judge pass/fail on benchmark basing on some arbitrary threshold during test execution.

So the ideal solution would be to distinguish between emasurements and results. Measurements should not have pass/fail associated with them as it makes little sense.

Patch-set: 2