Update patch set 2
Patch Set 2:

> I'm not sure which chart you're referring to,
The charts related to this boottime2 test are here:
https://validation.linaro.org/dashboard/image-charts/LCR-boot-trends
and https://validation.linaro.org/dashboard/image-charts/LCR-pre-merge
 > but for me this is a
 > major problem. Either the test result is a measurement (which btw
 > should not be associated with pass/fail value) or it's a pass/fail
 > result. 
for benchmark test, it should be recognized as pass if we could get measurement for it.
it's not necessary to have 2 entries(measurement and pass/faile result) for one test case
 > It's very inconvenient to alter between these basing on
 > measured value. I don't think it's a good idea to judge pass/fail
 > on benchmark basing on some arbitrary threshold during test
 > execution.

hmm, not sure why you get this image, the change here is to remove the measurement for a failed test case, there is no meaning to have measurement(the negative value here) for a failed test case.
 > 
 > So the ideal solution would be to distinguish between emasurements
 > and results.
agree on this.
 > Measurements should not have pass/fail associated with
 > them as it makes little sense.
partially agree here, but again from the implementation side, test cases with measurement should be considered as pass by default.

Patch-set: 2