Update patch set 6
Patch Set 6:

(2 comments)

When I started modifying the local output path, I found I have to modify almost all the __init__() functions of the classes. Because they do the args and test post processing separately and most of the lines do the same thing. I am moving those parts to main() to unify them, the idea is

* do args and test processing and variable assignment in main() and save the key:value pairs to test dictionary.
* pass args and test to classes and only set self.args and self.test in __init__()
* use self.args.name or self.test['key'] in classes

It was a mistake I made before, trying to fix it now, not sure if I am heading the right direction. Would you please give it a review?

Patch-set: 6
Label: Code-Review=0