Update patch set 2
Patch Set 2:

OK, let me ask from another side - what is the meaning of 'failed measurement'?

My point is that it makes hard to deal with the results on my end when it's not certain whether it's a 'pass/fail' or 'measurement' type of result. It's better to have a measurement of -1 as a sign of 'failed' test than nothing. This test is a benchmark. There should not be any failures here by definition.

Patch-set: 2