Update patch set 2
Patch Set 2:

If that boottime2 script fails to run correctly (does not detect required check-points), and detects that failure, it should not submit any measurements, because those will be bogus. 

I was prodding YongQin modify the boot time measurement script to do that. Not sure what the change really does, but what was there before (reporting test failed and returning a value of -1s?) was polluting trends with wrong measurements that we could have avoided.

Patch-set: 2