Update patch set 2
Patch Set 2:

> OK, let me ask from another side - what is the meaning of 'failed
 > measurement'?
I don't think there is meaning to have measurement for a failed test case.
we would not check the measurement of a failed test case I think.

 > 
 > My point is that it makes hard to deal with the results on my end
 > when it's not certain whether it's a 'pass/fail' or 'measurement'
 > type of result. 
Maybe I do not understand the user case you have to process here. 
 > It's better to have a measurement of -1 as a sign
 > of 'failed' test than nothing. This test is a benchmark. There
 > should not be any failures here by definition.
benchmark tests could also be failed, and even in some cases, some of the sub items succeed to get result, but some of the sub items failed to run to get result.
in such case, giving the failed test case a -1 as measurement is not good, something like NA(not numbers value) might be better, but the best is not to give them measurement I think.

Patch-set: 2