Update patch set 2
Patch Set 2:

YongQin - I'll repeat my question (again)- what does it mean that the benchmark is 'failed'? Does it mean that the value is below a certain threshold or sth else? Not giving a measurement to 'failed' benchmark converts it to 'pass/fail' type result. This is something I dislike more than not having the result recorded at all. To compare it to code it's like returning int or str from the same function depending on input value. This doesn't sound like a good practice.

Patch-set: 2